+++ src/index.ts
@@ -4,6 +4,7 @@
  */
 
 export * from './authorization';
+export * from './bulkDataAccess';
 export * from './bundle';
 export * from './capabilities';
 export * from './constants';
@@ -17,4 +18,6 @@ export * from './utilities';
 export * from './errors/InvalidResourceError';
 export * from './errors/ResourceNotFoundError';
 export * from './errors/ResourceVersionNotFoundError';
+export * from './errors/UnauthorizedError';
+export * from './errors/TooManyConcurrentExportRequestsError';
 export { stubs } from './stubs';