+++ src/stubs.ts
@@ -6,9 +6,10 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Search } from './search';
 import { History } from './history';
-import { Authorization } from './authorization';
+import { AccessBulkDataJobRequest, Authorization } from './authorization';
 import { Persistence } from './persistence';
 import { Bundle } from './bundle';
+import { BulkDataAccess } from './bulkDataAccess';
 
 export module stubs {
     export const bundle: Bundle = {
@@ -46,12 +47,21 @@ export module stubs {
     };
 
     export const passThroughAuthz: Authorization = {
-        async isAuthorized(request) {
-            return true;
+        // eslint-disable-next-line @typescript-eslint/no-empty-function
+        async isAuthorized(request) {},
+        // eslint-disable-next-line @typescript-eslint/no-empty-function
+        async isBundleRequestAuthorized(request) {},
+        async authorizeAndFilterReadResponse(request) {
+            return request.readResponse;
         },
+        // eslint-disable-next-line @typescript-eslint/no-empty-function
+        async isWriteRequestAuthorized(request) {},
 
-        async isBundleRequestAuthorized(request) {
-            return true;
+        // eslint-disable-next-line @typescript-eslint/no-empty-function
+        isAccessBulkDataJobAllowed(request: AccessBulkDataJobRequest) {},
+
+        getRequesterUserId(accessToken: string): string {
+            return 'random-userId';
         },
         async getAllowedResourceTypesForOperation(request) {
             return [
@@ -244,6 +254,20 @@ export module stubs {
 
         conditionalDeleteResource(request, queryParams) {
             throw new Error('Method not implemented.');
+        },
+    };
+
+    export const bulkDataAccess: BulkDataAccess = {
+        initiateExport(request) {
+            throw new Error('Method not implemented.');
+        },
+
+        cancelExport(jobId) {
+            throw new Error('Method not implemented');
+        },
+
+        getExportStatus(jobId) {
+            throw new Error('Method not implemented');
         },
     };
 }