+++ src/utilities.ts
@@ -4,6 +4,8 @@
  */
 
 import { TypeOperation, SystemOperation } from './constants';
+import { ExportType } from './bulkDataAccess';
+import { BulkDataAuth } from './authorization';
 
 export function chunkArray(myArray: any[], chunkSize: number): any[][] {
     const results = [];
@@ -50,9 +52,11 @@ export function getRequestInformation(
     resourceType?: string;
     id?: string;
     vid?: string;
+    bulkDataAuth?: BulkDataAuth;
 } {
     const path = cleanUrlPath(urlPath);
     const urlSplit = path.split('/');
+    const exportJobUrlRegExp = /\$export\/[\w|-]+/;
     switch (verb) {
         case 'PUT': {
             return {
@@ -69,6 +73,15 @@ export function getRequestInformation(
             };
         }
         case 'DELETE': {
+            if (exportJobUrlRegExp.test(urlPath)) {
+                const operation = 'cancel-export';
+                return {
+                    operation: 'delete',
+                    bulkDataAuth: {
+                        operation,
+                    },
+                };
+            }
             return {
                 operation: 'delete',
                 resourceType: urlSplit[0],
@@ -76,6 +89,32 @@ export function getRequestInformation(
             };
         }
         case 'GET': {
+            if (urlPath.includes('$export')) {
+                if (exportJobUrlRegExp.test(urlPath)) {
+                    const operation = 'get-status-export';
+                    return {
+                        operation: 'read',
+                        bulkDataAuth: {
+                            operation,
+                        },
+                    };
+                }
+
+                let exportType: ExportType = 'system';
+                if (urlPath.includes('/Patient/')) {
+                    exportType = 'patient';
+                }
+                if (urlPath.includes('/Group/')) {
+                    exportType = 'group';
+                }
+                return {
+                    operation: 'read',
+                    bulkDataAuth: {
+                        exportType,
+                        operation: 'initiate-export',
+                    },
+                };
+            }
             if (urlSplit[urlSplit.length - 1].startsWith('_history')) {
                 // if the last section of the url string starts with history
                 if (urlSplit[0].startsWith('_history')) {