+++ CHANGELOG.md
@@ -2,11 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+## [3.0.0] - 2020-11-11
+
+### Added
+- Required methods for SMART on FHIR interactions
+- Export UnauthorizedError to be usable
+- Required interfaces and types for bulk data export
+
 ## [2.0.0] - 2020-09-25
 
-- feat: Added `getAllowedResourceTypesForOperation` method to `Authorization` interface.
-- feat: Updated `Authorization.isAuthorized` to return a promise.
+### Added
+- `getAllowedResourceTypesForOperation` method to `Authorization` interface.
 
+### Updated
+- `Authorization.isAuthorized` to return a promise.
 
 ## [1.0.0] - 2020-08-31